requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.6
django-ckeditor==6.5.1
django-js-asset==2.0.0
idna==3.4
psycopg2-binary==2.9.5
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
